ml_pipeline/requirements.txt
# Numerical Computing
numpy>=1.26.0

# Data Manipulation
pandas>=2.0.0

# Database Connectivity
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Configuration
PyYAML>=6.0.0